:construction: start wars³     - get forward and right player vector direction     - shoot ray to walls     - refactor cub_get_mouse_position fix :     - save game serialization

+++ src/io/parsing_utils.c
@@ -72,8 +72,8 @@ void	add_map_row(t_map_info *infos, char *line)
 		infos->map_raw = gc_strarray_init();
 	infos->map_raw = gc_strarray_append(infos->map_raw, line);
 	line = ft_trimr(line);
-	if (ft_strlen(line) > infos->width)
-		infos->width = ft_strlen(line);
+	if ((int64_t)ft_strlen(line) > infos->width)
+		infos->width = (int64_t)ft_strlen(line);
 }
 
 static int	contains_illegal_char(t_map_info *infos, char *line)
